helios-testing: Address PR comments
